data-processing/scripts/etl/requirements.txt
certifi==2026.4.22
influxdb-client==1.50.0
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reactivex==4.1.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
